app/requirements.txt
aniso8601==8.1.1
attrs==20.3.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
flasgger==0.9.5
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
imageio==2.9.0
imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
networkx==2.5
numpy==1.19.4
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
Pillow==8.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scipy==1.6.0
six==1.15.0
tifffile==2021.1.14
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
